Removed custom serialization support
