+++ src/main/java/org/opensearch/security/auditlog/impl/AbstractAuditLog.java
@@ -775,8 +775,7 @@ public abstract class AbstractAuditLog implements AuditLog {
         if (address == null && threadPool.getThreadContext().getHeader(ConfigConstants.OPENDISTRO_SECURITY_REMOTE_ADDRESS_HEADER) != null) {
             address = new TransportAddress(
                 (InetSocketAddress) Base64Helper.deserializeObject(
-                    threadPool.getThreadContext().getHeader(ConfigConstants.OPENDISTRO_SECURITY_REMOTE_ADDRESS_HEADER),
-                    threadPool.getThreadContext().getTransient(ConfigConstants.USE_JDK_SERIALIZATION)
+                    threadPool.getThreadContext().getHeader(ConfigConstants.OPENDISTRO_SECURITY_REMOTE_ADDRESS_HEADER)
                 )
             );
         }
@@ -787,8 +786,7 @@ public abstract class AbstractAuditLog implements AuditLog {
         User user = threadPool.getThreadContext().getTransient(ConfigConstants.OPENDISTRO_SECURITY_USER);
         if (user == null && threadPool.getThreadContext().getHeader(ConfigConstants.OPENDISTRO_SECURITY_USER_HEADER) != null) {
             user = (User) Base64Helper.deserializeObject(
-                threadPool.getThreadContext().getHeader(ConfigConstants.OPENDISTRO_SECURITY_USER_HEADER),
-                threadPool.getThreadContext().getTransient(ConfigConstants.USE_JDK_SERIALIZATION)
+                threadPool.getThreadContext().getHeader(ConfigConstants.OPENDISTRO_SECURITY_USER_HEADER)
             );
         }
         return user == null ? null : user.getName();
