+++ src/main/java/org/opensearch/security/transport/SecurityInterceptor.java
@@ -39,7 +39,6 @@ import org.apache.commons.lang3.StringUtils;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 
-import org.opensearch.Version;
 import org.opensearch.action.admin.cluster.shards.ClusterSearchShardsAction;
 import org.opensearch.action.admin.cluster.shards.ClusterSearchShardsResponse;
 import org.opensearch.action.get.GetRequest;
@@ -64,7 +63,6 @@ import org.opensearch.security.ssl.transport.SSLConfig;
 import org.opensearch.security.support.Base64Helper;
 import org.opensearch.security.support.ConfigConstants;
 import org.opensearch.security.support.HeaderHelper;
-import org.opensearch.security.support.SerializationFormat;
 import org.opensearch.security.user.User;
 import org.opensearch.threadpool.ThreadPool;
 import org.opensearch.transport.Transport.Connection;
@@ -155,7 +153,6 @@ public class SecurityInterceptor {
 
         final boolean isDebugEnabled = log.isDebugEnabled();
 
-        final var serializationFormat = SerializationFormat.determineFormat(connection.getVersion());
         final boolean isSameNodeRequest = localNode != null && localNode.equals(connection.getNode());
 
         try (ThreadContext.StoredContext stashedContext = getThreadContext().stashContext()) {
@@ -238,37 +235,18 @@ public class SecurityInterceptor {
             }
 
             try {
-                if (clusterInfoHolder.getMinNodeVersion() == null || clusterInfoHolder.getMinNodeVersion().before(Version.V_2_14_0)) {
-                    if (serializationFormat == SerializationFormat.JDK) {
-                        Map<String, String> jdkSerializedHeaders = new HashMap<>();
-                        HeaderHelper.getAllSerializedHeaderNames()
-                            .stream()
-                            .filter(k -> headerMap.get(k) != null)
-                            .forEach(k -> jdkSerializedHeaders.put(k, Base64Helper.ensureJDKSerialized(headerMap.get(k))));
-                        headerMap.putAll(jdkSerializedHeaders);
-                    } else if (serializationFormat == SerializationFormat.CustomSerializer_2_11) {
-                        Map<String, String> customSerializedHeaders = new HashMap<>();
-                        HeaderHelper.getAllSerializedHeaderNames()
-                            .stream()
-                            .filter(k -> headerMap.get(k) != null)
-                            .forEach(k -> customSerializedHeaders.put(k, Base64Helper.ensureCustomSerialized(headerMap.get(k))));
-                        headerMap.putAll(customSerializedHeaders);
-                    }
-                }
+                Map<String, String> jdkSerializedHeaders = new HashMap<>();
+                HeaderHelper.getAllSerializedHeaderNames()
+                    .stream()
+                    .filter(k -> headerMap.get(k) != null)
+                    .forEach(k -> jdkSerializedHeaders.put(k, headerMap.get(k)));
+                headerMap.putAll(jdkSerializedHeaders);
                 getThreadContext().putHeader(headerMap);
             } catch (IllegalArgumentException iae) {
                 log.debug("Failed to add headers information onto on thread context", iae);
             }
 
-            ensureCorrectHeaders(
-                remoteAddress0,
-                user0,
-                origin0,
-                injectedUserString,
-                injectedRolesString,
-                isSameNodeRequest,
-                serializationFormat
-            );
+            ensureCorrectHeaders(remoteAddress0, user0, origin0, injectedUserString, injectedRolesString, isSameNodeRequest);
 
             if (actionTraceEnabled.get()) {
                 getThreadContext().putHeader(
@@ -295,8 +273,7 @@ public class SecurityInterceptor {
         final String origin,
         final String injectedUserString,
         final String injectedRolesString,
-        final boolean isSameNodeRequest,
-        final SerializationFormat format
+        final boolean isSameNodeRequest
     ) {
         // keep original address
 
@@ -334,11 +311,10 @@ public class SecurityInterceptor {
                 getThreadContext().putTransient(ConfigConstants.OPENDISTRO_SECURITY_INJECTED_USER, injectedUserString);
             }
         } else {
-            final var useJDKSerialization = format == SerializationFormat.JDK;
             if (transportAddress != null) {
                 getThreadContext().putHeader(
                     ConfigConstants.OPENDISTRO_SECURITY_REMOTE_ADDRESS_HEADER,
-                    Base64Helper.serializeObject(transportAddress.address(), useJDKSerialization)
+                    Base64Helper.serializeObject(transportAddress.address())
                 );
             }
 
@@ -346,10 +322,7 @@ public class SecurityInterceptor {
             if (userHeader == null) {
                 // put as headers for other requests
                 if (origUser != null) {
-                    getThreadContext().putHeader(
-                        ConfigConstants.OPENDISTRO_SECURITY_USER_HEADER,
-                        Base64Helper.serializeObject(origUser, useJDKSerialization)
-                    );
+                    getThreadContext().putHeader(ConfigConstants.OPENDISTRO_SECURITY_USER_HEADER, Base64Helper.serializeObject(origUser));
                 } else if (StringUtils.isNotEmpty(injectedRolesString)) {
                     getThreadContext().putHeader(ConfigConstants.OPENDISTRO_SECURITY_INJECTED_ROLES_HEADER, injectedRolesString);
                 } else if (StringUtils.isNotEmpty(injectedUserString)) {
