+++ src/main/java/org/opensearch/security/transport/SecurityRequestHandler.java
@@ -107,8 +107,6 @@ public class SecurityRequestHandler<T extends TransportRequest> extends Security
             resolvedActionClass = ((ConcreteShardRequest<?>) request).getRequest().getClass().getSimpleName();
         }
 
-        final boolean useJDKSerialization = getThreadContext().getTransient(ConfigConstants.USE_JDK_SERIALIZATION);
-
         String initialActionClassValue = getThreadContext().getHeader(ConfigConstants.OPENDISTRO_SECURITY_INITIAL_ACTION_CLASS_HEADER);
 
         final ThreadContext.StoredContext sgContext = getThreadContext().newStoredContext(false);
@@ -183,7 +181,7 @@ public class SecurityRequestHandler<T extends TransportRequest> extends Security
                 } else {
                     getThreadContext().putTransient(
                         ConfigConstants.OPENDISTRO_SECURITY_USER,
-                        Objects.requireNonNull((User) Base64Helper.deserializeObject(userHeader, useJDKSerialization))
+                        Objects.requireNonNull((User) Base64Helper.deserializeObject(userHeader))
                     );
                 }
 
@@ -192,7 +190,7 @@ public class SecurityRequestHandler<T extends TransportRequest> extends Security
                 if (!Strings.isNullOrEmpty(originalRemoteAddress)) {
                     getThreadContext().putTransient(
                         ConfigConstants.OPENDISTRO_SECURITY_REMOTE_ADDRESS,
-                        new TransportAddress((InetSocketAddress) Base64Helper.deserializeObject(originalRemoteAddress, useJDKSerialization))
+                        new TransportAddress((InetSocketAddress) Base64Helper.deserializeObject(originalRemoteAddress))
                     );
                 } else {
                     getThreadContext().putTransient(ConfigConstants.OPENDISTRO_SECURITY_REMOTE_ADDRESS, request.remoteAddress());
